requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# LLM
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
langchain-core==0.1.10
openai>=1.10.0

# Document Processing
pypdf==3.17.0
tiktoken==0.5.2

# Vector Store
faiss-cpu>=1.8.0

# Utils
python-dotenv==1.0.0
tenacity==8.5.0
httpx==0.26.0

# AWS
boto3==1.34.0
